chore: satisfy parcel watcher repository gates
