+++ workspace-architecture.json
@@ -25,7 +25,7 @@
     ]
   },
   "baseline": {
-    "workspace_members": 78,
+    "workspace_members": 79,
     "default_dependency_closure": [
       "perry",
       "perry-api-manifest",
@@ -66,7 +66,7 @@
       "perry-updater"
     ],
     "decision_counts": {
-      "externalize": 31,
+      "externalize": 32,
       "keep": 42,
       "merge": 1,
       "remove": 1,
@@ -203,6 +203,10 @@
       "decision": "externalize"
     },
     "perry-ext-node-forge": {
+      "category": "binding",
+      "decision": "externalize"
+    },
+    "perry-ext-parcel-watcher": {
       "category": "binding",
       "decision": "externalize"
     },
